Introduce `Driver.getServerInfo` method
This is part of the drivers unification initiative.

The `Driver.verifyConnectivity` method was implemented in different ways accross the drivers with
different returns value. For concistency and clarity about what the method does, the new behaviour
of the `Driver.verifyConnectivy` is check the connectivity with all the readers available and then
returns a resolve empty promise in case of success (the promise will be rejected, otherwise).

The Server Info information was moved to the method `Driver.getSeverInfo` which verifies the
connectivity and then returnt the `ServerInfo` of one of the read servers.

For keeping the backwards compatibility, the `Driver.verifyConnectivity` was depreacted in this version
and it should be replace with a version with `Promise<void>` as return value in the next major release.
`Driver.getServerInfo` should be used instead if the client code needs server information.