Introduce `Driver.getServerInfo` method (#878)
This is part of the drivers unification initiative.

The `Driver.verifyConnectivity` method was implemented in different ways across the drivers with different returns value. For consistency and clarity about what the method does, the new behaviour of the `Driver.verifyConnectivity` is check the connectivity with all the readers available and then returns a resolve empty promise in case of success (the promise will be rejected, otherwise).

The Server Info information was moved to the method `Driver.getSeverInfo` which verifies the connectivity and then returns the `ServerInfo` of one of the read servers.

For keeping the backwards compatibility, the `Driver.verifyConnectivity` was deprecated in this version and it should be replace with a version with `Promise<void>` as return value in the next major release. `Driver.getServerInfo` should be used instead if the client code needs the server information.