Fix `TypeError: Cannot read properties of undefined (reading 'isActive')` (#881)
This error occurs in situations where the pool was already close, the acquire resource was already completed and a pool was recreated in the meantime. In this situation,
the _release method is called again but with an empty pool state causing the error.

For solving this issue, we should check if the state is defined before check the `isActive`. Undefined poolState should be considered closed in this case and the connection
should be detroyed.