Improve back-pressure mechanism for RxResult (#882)
The lack of consumer oriented back-pressure mechanism in the RxJS apis was causing memory issues and other problems in the client code. 

For solving this issue, the records observer returned by `RxResult.records()` was changed for using the async iterator as foundation. New methods for enabling the client for fine controlling the stream were added to the `RxResult` api. These methods are:

- `pause()`: Pause the record streaming. No new record will be pushed to the stream util `push` or `resume` get called. 
- `resume()`: Resumes the records streaming. 
- `push()`: Push the next record. If the stream is not paused, this method will pause it for giving the push control to the client.