Avoid to send reset when it is not needed
The `RESET` should be send when a failure occurs or whenever the connection is being sent back
to the pool with a pending request running, i.e. the bolt server is not in the `READY` state.

These changes also affect the `verifyConnectivity` and `getServerInfo` implementation.
The `RESET` message is not sent in these methods if it is a newly created connection.