Skip flaky test: test_should_write_successfully_on_leader_switch_using_tx_function (#905)
This test fails because of Javascript driver doesn't validate the tx before send it to the tx function.

In this case, the transaction could be in a failed state when `tx.run` get called and it produces the
`DriverError :  : Cannot run query in this transaction, because it has been rolled back either because of an error or explicit termination. (N/A)`.

This errant behaviour of the driver was fixed in the 5.0 with the TransactionPromise.