Standardise code and fix linter issues (#919)
Fix and apply the configured standard to the code base makes easier for the next changes go through the pipeline without have to mess up with tech debts. This also enables the lint stage be part of test routine, not only in the pre-commit hook (which could be by passed).