Refactory `Pool` module
The state of the pool and the resource list weren't bounded togheter as it supposed to. This scenario makes the code prune to bugs and incosistency in the long run.

Merge the resouse list (a.k.a. Pool) and PoolState in the same object makes the code easier to understand and more coesive.