Keep seed router address in the pool after routing table
Forgetting the seed router was causing the connection to the seed router being recreated whenever the routing table is fetched and the seed router is not part of the cluster formation. For cases where the database name is set in the session, the driver will have to fetch the routing table for the default/home database for discovering the database name and make it consitent during the session lifetime. In this scenario, the seed router will be always used during the discovery. So, a new connection to the seed router will be created for each session.

The solution for this issue is to keep the seed router in the pool during the update routing table process. This way, the same connections for the seed router can be re-used.