Refactory `Pool` module (#930)
The state of the pool and the resource list weren't bounded together as it supposed to. This scenario makes the code prune to bugs and inconsistency in the long run.

Merge the resource list (a.k.a. Pool) and PoolState in the same object makes the code easier to understand and more cohesive.