Decouple Packstream Serialization/De-serialization from Bolt structut hydration
The current implementation of the packstream packer and unpacker also includes the creation of the bolt structures. This implementation join twos different concepts in the same object and this doesn't allow the flexibility needed for working with different structure for the same object in the same bolt version, for instance.

Decoupling this two layer allows the driver supports the new `utc` feature in a more suitable way.