Validate the ZoneId of the DateTime with ZoneId
The validation of the DateTime was only being done in the new patched driver while unpacking the struct. This changes force any new DateTime with ZoneID to have a valid ZoneId.

This changes also treats struct unpacking errors and deffers the occurred to the moment the object is manipulate. For instance, a DateTime with invalid ZoneId returned in a Record
will not break the records consumption until any code try to interacts with the broken DateTime.