Add property-based testing to temporal-types conversion
Add this type of testing to `.toStandardDate()` ('should be the reverse operation of fromStandardDate but losing time information') helps to cover corner cases and solve special cases such:

* Negative date time not being serialized correctly in the iso standard. Years should always have 6 digits and the signal in front for working correctly with negative years and high numbers.
This also avoids the year 2000 problem. See, https://en.wikipedia.org/wiki/ISO_8601
* `Date.fromStandardDate` factory was not taking in consideration the `seconds` contribuition in the timezone offset.
This is not a quite common scenarion, but there are dates with timezone offset of for example `50 minutes` and `20 seconds`.
* Fix `Date.toStandardDate` for dates with offsets of seconds.
Javascript Date contructor doesn't create dates from iso strings with seconds in the offset.
For instance, `new Date("2010-01-12T14:44:53+00:00:10")`.
So, the date should be re-created from the utc timestamp.