Add JSON serialization methods for Date types
This enables `JSON.stringify` to serialize these data types to ISO strings instead of JS Objects.
Transformation loops and other bollerplates like the following with this method introduction.

For instance, the following code:

```javascript
if (result.records) {
    for (const record of result.records) {
        for (const recordKey of record.keys) {
            const object = record.get(recordKey);
            if (!Array.isArray(object)) {
                for (const key in object) {
                    if (object.hasOwnProperty(key) && isDateTime(object[key])) {
                        object[key] = object[key].toString();
                    }
                }
            }
        }
    }
}
```
This snippet can be replaced by this other one:

```javascript
if (result.records) {
    for (const record of result.records) {
      const object = record.toObject()
    }
}
```