Commit `neo4j-driver-deno` to the repository
The `deno` driver is generated using a custom script made for this repository.
This script doesn't have any tests linked to it.
So the only way this changes are tested is by running the `deno driver` test suite.
For guarantee the behaviour of this driver, we should version commit the generated driver to the repository.

This changes introduce the first commited version of the driver.
Adding the verification if the commited deno driver is in sync with the lite driver is also part of the this scope.