Fail fast on rediscovery when 'Neo.ClientError.Statement.ArgumentError' failure
When trying to impersonate an invalid user, the routing procedure returns `Neo.ClientError.Statement.ArgumentError`.
This type of failure is not in the fail fast list, thus this is wrapped up in the generic rediscovery error triggering the retry in the transaction functions.
Since this kind of error is not recoverable, retrying on it is not need and desirable.

Then, failing fast on `Neo.ClientError.Statement.ArgumentError` failures speeds up the error bubbling and avoids unnecessary load in the client and server.