Clear receive timeout when close browser and Deno channels (#1016)
The receive timeout was not being clearer when channel get closed.
With the timeout running, receive timeouts events can still be notified to the existing observers.
This was causing some tests failing and possible issues in production code.

This bug was found in the test `package/neo4j-driver/test/result.test.js (should handle missing onCompleted)`. This test was calling `done` twice because of a late failing coming to the observer.

Node channels are not affected by this error since its timeout implementation is controlled to the socket.