Enables Deno specific tests in Core
Core features which are environment dependent will need to be implemented.
For testing this features, environment dependent tests need to be setup.

The `deno` tests can be run by `npm run test::deno` and they need to be located under `./packages/core/test/deno`.
The testkit pipeline is configured to run these tests only when configured for test `deno`.