Fix unit/integration tests related with connection.connect and bolt-agent for Deno
The test code related with `connection.connect` were calling the method without the bolt agent causing all kind of errors.
Add the bolt agent  to the calls and expectations solves the issue.

`Deno.osRelease()` and the flag `--allow-sys` are not available in Deno 1.19.2.
The missing flag was causing the test don't even start.
The method unexistance was causing the code fails because is calling a non existing method.

The missing flag is solved by using `--allow-all` instead.
This is a more permissive flag, but it is not a big deal in test env.

The absence of `Deno.osRelease()` is solved by not setting the release when this method is not available.
This might be an issue for old Deno users, but it's not a blocker.