Remove `boltAgent` from API and fix declaration tests
`boltAgent` is not configurable, so it should not be exposed in the `Config`.
The `InternalConfig` interface was created for keeping the type-safety.

The declaration tests were failing because of multiple `AbortSignal` definitons.
This was happening because each project is using different versions of `@types/node` in the `package-lock.json`.
The problem was not happening before because any interface exposed by the `neo4j-driver-core` was depending in Node defintions.
The solution can be done or by remove the Node api  reference or by pin the  `@types/node` versions accross all packages.
The first solution was choosed since the use of `string` instead `NodeJS.Platform` in the `SystemInfo` doesn't impact the usability of the `node.boltAgent.fromVersion` function.