Fix browser test in Node10 environment
The window mock was broken because it miss `Symbol`, `Date` and other global definitions.
Set the window as `globalThis` and then define the navigator in it solves the issue.

A proper way of solving the issue will be use `jsdom`, however this plugin doesn't work in Node10.

Also fixed an issue with timers getting hanging in some situations. The issue was catched by browser test with `--detectOpenHandles`.