Fix browser `boltAgent` in web workers context (#1094)
The driver was trying to access the `window` object inside web workers. This object doesn't exist in this context.
So, the driver was throwing an error and not creating connections at all.

This problem is solved by access the `navigator` directly from the global scope (which is the `window` outside the web workers). Some guards were added to avoid the code break in case `navigator` is not defined.