Add `notificationFilter` to the Config type (#1093)
The `notificationFilter` config was not mapped correctly to the `Config` type. So, the driver was not offering the correct auto-complete and type support for this configuration on the `neo4j.driver` factory.

The disconnection between the configuration object and the documentation was the root cause of the issue, since the filter was correctly documented.

For solving this issue, the `Config` documentation was moved to live along with its definition. The interface was promoted to class to make it happen. However, the class constructor can only be used internally and for the api point of view, the `Config` still behaving as an interface.