Decrease integration tests duration
80% of the integration tests was related to random temporal types tests.
Make this tests run in parallel by starting all of them using `setImmediate` function.

Since `setImmediate` is not available in browsers, this tests will be skipped on this environment.
This is not a issue since these tests paths are already cover by property tests in the `bolt-connection` unit tests suite
and in Testkit.