Fix DateTime with ZoneId unpacking
The timezone offset was miss-calculated because of an error on extracting timezone information when the hour equals to `0`.
The problems happens because `Intl.DateTimeFormat` when configured with `hour12: false` returns `0` hour as `24`.

The solution for this is convert `24` to `0` before calculate the `offset`.

NOTE:
Other valid solution would be change `hourCycle` to `h23`.
However, this solution is not supported by all javascript environment.