Export `RecordShape` type
This type is the base type used to define the shape of the Record.
Exporting this type enables user to use it for defining custom generic function to work with result and records without have to redefine the shape type.

The type was renamed from `Dict` to `RecordShape` for having a more meaniful name for external consumers.