Fix flaky tests on TransactionExecutor suite
The TransactionExecutor tests were mocking the global setTimeout and clearTimeout.
This mocks were conflicting with other calls to this api in the browser tests.
So, the calls to the set/clear were not being the expected in some situations.

Replace the global mock for injecting the functions resolve this issue in a less intrusive way.