Avoid Results try to pull more records after transaction fail
When two or more queries are running concurrent in the same transaction, a failure happenning in any of the queries implies in a broken transaction and any new message to the server should result in a error since the server state will be in failure.

Notifying all open results in the transaction about any error avoids any new message to be send by these objects.