Introduce explicity resource management to Driver and Session objects
This is a TC39 [proposal](https://github.com/tc39/proposal-explicit-resource-management) which is already implemented in Typescript 5.2, core-js, babel and other pollifyl tools.

This feature enables the user create the driver or a session with the `await using` keywords and then do not have to close the resource afterwards, since this resources will be closed after leaving the block which were created at.

For example:

```typescript
await using driver = neo4j.driver(uri, authToken)
await using session = driver.session()

await session.executeRead(tx => "RETURN 1")
```

Since Deno is more strict with typescript, small fixes had to be done in the driver.

* Add a value to Result[Symbol.toStringTag].
* Fix Integer.shiftRight function to add propert integer convertion it.

A package.json file was added to the `neo4j-driver-deno` folder for making easier to integrate the test runners to the environments.