Fix BrowserChannel close when is already closing
WebSockets can take a while for closing and the `WebSocket.close` is
called when status is `WS_CLOSING`, an error is thrown.

This situation can happen when the driver is being closed while session
still releasing connections back to the pool or after a receive timeout
while closing the session.

BrowserChannel should wait for the original close socket finish whenever
a new close request happens to avoid this kind of error and provides a
graceful shutdown for driver and session.