Fix running into an infinite loop when closing a transaction with still running queries.
When closing a transaction, that still had queries open the `_onErrorCallback` in core/src/transaction.ts:303 called
`resultStreamObserver.onError` for each result. The default behaviour for `FailedObvserver` (one implementation) is to call `this._beforeError`(core/src/internal/observer.ts:179), which is set to `onError`, which happens to be set to the mentioned
`_onErrorCallback`.
This results in an infinite async-loop, which gradualy consumes all available memory and crashes the process.

Because the `FAILED` state is only set in `_onErrorCallback` we can use it as a flag, to cut off the infinite recursion, when we are allready in the `FAILED` state.