Introduce `GqlStatusObject` support as notifications to `ResultSummary` (#1194)
**⚠️ This API is released as preview.**

Introduces `ResultSummary.gqlStatusObjects` as preview feature. 

`GqlStatusObject` is a GQL compliant `Notification` object and status of query execution, this new object includes `gqlStatus` and `statusDescription`.

`ResultSummary.gqlStatusObjects` always contains at least 1 status representing the `Success`, `No Data` or `Omitted Result`. When discarding records in a RxSession while connected to a non-GQL-aware server, the driver might not be able to tell apart `Success` and `No Data`.

The GqlStatusObjects will be presented in the following order:

* A “no data” (`02xxx`) has precedence over a warning;
* A warning (`01xxx`) has precedence over a success.
* A success (`00xxx`) has precedence over anything informational (`03xxx`)

### Migrating from Notification

Most of the properties present in the `Notification` are present in the new object, except `code` and `title`. 
The `GqlStatusObject.gqlStatus` is equivalent of `code` in the `Notification`, but with values compliant with GQL.
The properties `GqlStatusObject.classification` and `GqlStatusObject.rawClassification` are equivalent to `Notification.category` and `Notification.rawCategory`. The name change is needed because `category` has a different meaning in GQL.

### Configuration
Filtering `gqlStatusObjects` can be done using in the same way as filtering `notifications`. However, non-notification status can not be filtered (such as `Success` or `No Data`).
The property `disabledClassifications` was added to `NotificationFilter` for being named consistently with GQL. This property is equivalent to `disabledCategories` and it be can used interchangeably. However, they can not be used at the same time. An exception will be raised in this scenario.   

**⚠️ This API is released as preview.**